requirements.txt
requests==2.31.0
instagrapi==2.0.0
schedule==1.2.0
python-dotenv==1.0.0
Pillow==10.1.0
feedparser==6.0.10
beautifulsoup4==4.12.2
openai==1.3.0
newspaper3k==0.2.8
selenium==4.15.2
webdriver-manager==4.0.1
